youtube.Search: Implement uppercase & symbols modes; tvOS 14 look
The `stbt.Keyboard` API has changed in Stb-tester v32. The internal
representation of the Directed Graph changed and is now private, so we
can't manipulate it directly with `networkx.compose_all`. Instead, we
create the Keyboard instance: `kb = stbt.Keyboard()` and then add to the
model incrementally by calling `kb.add_grid` and `kb.add_transition` to
specify all the keys & transitions.

At the same time we have updated the reference images to match the
latest appearance of the YouTube app on tvOS 14.

For a detailed description of this code, see the tutorial at
https://stb-tester.com/blog/2020/10/02/testing-onscreen-keyboards